partie client

+++ src/main/java/spring/devoir/gestion_tickets/repositories/TicketRepository.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface TicketRepository extends JpaRepository<Ticket, Long> {
 
     List<Ticket> findByDeveloppeurIsNull();
-    List<Ticket> findByDeveloppeur_Id(int id);
-    List<Ticket> findByClient_Id(int id);
+    List<Ticket> findByDeveloppeur_Id(Long id);
+    List<Ticket> findByClient_Id(Long id);
 }
